+++ app/src/main/java/com/example/newsapp/ui/MainViewModel.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.newsapp.domain.usecases.AppEntryUseCases
+import com.example.newsapp.domain.usecases.appentry.AppEntryUseCases
 import com.example.newsapp.ui.navgraph.Route
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.delay
